Add transit data caching
Starting up with the cache is about 5x faster than without on my
machine, at the cost of additional time for serializing the cache if it
does not already exist. Specifically in my testing ~15 seconds startup
became ~22s first startup and ~3s cached startup.

This necessitates removing transitfeed's StopTime class from
transit_times because attempting to pickle its StopTime fails. When
using protocol 0 or 1 it results in

    TypeError: a class that defines __slots__ without defining __getstate__ cannot be pickled

and indeed StopTime defines the former and not the latter.

When using protocol 2 it results in

    pickle.PicklingError: Can't pickle <function <lambda> at 0x7f1b00cdcc80>: it's not found as transitfeed.schedule.<lambda>

which seems to be because a StopTime includes a Stop, which includes a
Schedule, which sets a field to `defaultdict(lambda: [])`.